refactor: move categories to external component Categories.js
closed #107
